requirements.txt
# Core vision pipeline
opencv-python>=4.8.0
numpy>=1.24.0

# Unity asset extraction
UnityPy>=1.10.0
Pillow>=10.0.0

# OCR
pytesseract>=0.3.10

# ML
ultralytics>=8.0.0
lightgbm>=4.0.0

# Screen capture
mss>=9.0.0

# UI (uncomment chosen approach at implementation time)
# PyQt6>=6.5.0
# flask>=3.0.0